Add scrollLock back to useScrollTo for mobile. Also add/move the desktop scroll lock (scroll.stop).
I don't know why stopping the scroll didn't stop the scrolling in the new view earlier. When moved it to the hook, scroll was stopped in the new view, so that's why there is now scroll.start(); which how it should've been always! 🤷‍♂️

Scroll.update now unnecessary since there are nothing to update (e.g. multiple pages at the same time which needed updates or match errors we're catched).

Note: We could allow the page transitions now without errors and without locking the scroll (needs scroll.update() in App.js) but it's kind of pointless because there is no content to show until the transition finishes and new page loads.