Handle possible layout shift w/ scroll bar width
Handling scroll bar hiding like this makes the wrap work correctly in case the browser has visible scroll bars. Earlier when using 100vw, some of the content was hidden under the scroll bar and made the right wrap smaller than it should be.

Downside is that when the overflow is hidden, the :root background leaks under the Template background and could create a visual inconsistency with the current page's theme.

Note that this change is only visible if the browser has visible scroll bars (e.g. not overlayed ones).

Also, remove the unnecessary width calculation from Milestones' line chart.